Sentiment-Analysis-Chatbot/requirements.txt
langchain
langchain-community
langchain-groq



python-dotenv
